[4.5][ARCH] Complete A4 micro-debt: user_env split and v1 contract warn.
Split host/user_env into parse and wizard modules; warn when flat v1
odpm.json omits odpm_version; update architecture debt tracker and docs.
