+++ dev_project/constants/scenarios.py
@@ -13,7 +13,7 @@ VENV_MODE_FRESH = "fresh"
 VENV_MODE_BAKED = "baked"
 VENV_MODE_VALUES = frozenset((VENV_MODE_FRESH, VENV_MODE_BAKED))
 
-DEFAULT_ODPM_VERSION = "3.0"
+DEFAULT_ODPM_VERSION = "3.0"  # legacy v1 fallback when odpm_version is missing; see manifest/v1_contract.py
 # Single user-facing product version: CLI, pip/PyPI, deb/rpm, git tag (v{RELEASE_VERSION}).
 RELEASE_VERSION = "4.4.3"
 # Recommended stable line for install docs and mike `stable` alias (bump on stable tag only).
